UserInterface
Wrote a primitive userIntface that is probably a bit heavy on the code. I also have it almost all in one method since it is a lot of typing and responding, so maybe we should change that. I have tested the userinterface a lot though. Also had to change moss a little bit so it played nice with a user interface